requirements.txt
django
django-distill
django-cleanup
django-solo
sorl-thumbnail
Pillow
markdown
more_itertools
pyyaml
django-summernote